feat: enhance evoucher functionality with new schema and public endpoints
- Added EvoucherCode schema and integrated it into the evoucher module.
- Implemented public GET endpoint to retrieve available evouchers for users.
- Refactored evoucher code service to validate evoucher expiration and user claims.
- Cleaned up unauthorized access checks in the evoucher code controller.
- Updated utility functions for better validation of evoucher availability and claims.
