+++ backend/app/src/module/evoucher/evoucher.module.ts
@@ -5,6 +5,8 @@ import { Sponsors } from '../sponsors/schema/sponsors.schema';
 import { SponsorsSchema } from '../sponsors/schema/sponsors.schema';
 import { EvoucherController } from './controller/evoucher.controller';
 import { EvoucherService } from './service/evoucher.service';
+import { EvoucherCode } from './schema/evoucher-code.schema';
+import { EvoucherCodeSchema } from './schema/evoucher-code.schema';
 
 @Module({
   imports: [
@@ -17,6 +19,10 @@ import { EvoucherService } from './service/evoucher.service';
         name: Sponsors.name,
         schema: SponsorsSchema
       },
+      {
+        name: EvoucherCode.name,
+        schema: EvoucherCodeSchema
+      }
     ])
   ],
   exports: [MongooseModule],
